Endrer forventet return type for count() fra BigInteger til Long (#8726)

+++ domenetjenester/behandlingskontroll/src/main/java/no/nav/k9/sak/behandlingskontroll/impl/BehandlingStegVisitor.java
@@ -313,7 +313,7 @@ class BehandlingStegVisitor {
 
     private void validerTilbakeføringUtenAksjonspunktCircuitBreaker(Behandling behandling) {
         LocalDateTime ettDøgnSiden = LocalDateTime.now().minusDays(1);
-        int antallTilbakeføringer = behandlingRepository.antallTilbakeføringerSiden(behandling.getId(), ettDøgnSiden);
+        long antallTilbakeføringer = behandlingRepository.antallTilbakeføringerSiden(behandling.getId(), ettDøgnSiden);
 
         if (antallTilbakeføringer > 100) {
             throw new IllegalStateException("Mulig evig løkke ved tilbakeføring. Har hatt " + antallTilbakeføringer + " tilbakeføringer uten aksjonspunkt for behandlingen siden " + ettDøgnSiden + ". Stopper prosessering midlertidig. ");
